+++ servers/Azure.Mcp.Server/CHANGELOG.md
@@ -6,6 +6,8 @@ The Azure MCP Server updates automatically by default whenever a new release com
 
 ### Features Added
 
+- Replace hard-coded strings for Azure.Mcp.Server with ones from IConfiguration. [[#1269](https://github.com/microsoft/mcp/pull/1269)]
+
 ### Breaking Changes
 
 ### Bugs Fixed
@@ -61,6 +63,7 @@ The Azure MCP Server updates automatically by default whenever a new release com
 ### Other Changes
 
 - Added a `CancellationToken` parameter to async methods to more `I[SomeService]` interfaces. [[#1178](https://github.com/microsoft/mcp/pull/1178)]
+- Update server to support official MCP repository. [[#1104](https://github.com/microsoft/mcp/pull/1104)]
 
 ## 2.0.0-beta.4 (2025-11-13)